Add files via upload
In this project, we are analyzing different Warby Parker’s marketing funnels in order to calculate conversion rates. 

Here are the funnels and the tables that were given:

Quiz Funnel: - survey

Home Try-On Funnel: 
   - quiz
   - home_try_on
   - purchase

This project is a collaboration with Warby Parker’s Data Science team and uses fictional data.